+++ scripts/rollback/rollback-3.1.6.sh
@@ -1,0 +1,296 @@
+#!/bin/bash
+# timestamp: 1647856682908
+
+npm dist-tag add @midwayjs/egg-layer@2.14.7 latest
+npm dist-tag add @midwayjs/egg-layer@3.1.6 beta
+npm dist-tag add @midwayjs/express-layer@3.1.5 latest
+npm dist-tag add @midwayjs/express-layer@3.1.6 beta
+npm dist-tag add @midwayjs/faas-typings@3.1.5 latest
+npm dist-tag add @midwayjs/faas-typings@3.1.6 beta
+npm dist-tag add @midwayjs/koa-layer@3.1.5 latest
+npm dist-tag add @midwayjs/koa-layer@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-app@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-app@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-aws-starter@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-fc-starter@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-fc-trigger@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-http-parser@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-http-parser@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-scf-starter@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-scf-trigger@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-vercel-starter@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@3.1.6 beta
+npm dist-tag add @midwayjs/serverless-worker-starter@3.1.5 latest
+npm dist-tag add @midwayjs/serverless-worker-starter@3.1.6 beta
+npm dist-tag add @midwayjs/static-layer@3.1.5 latest
+npm dist-tag add @midwayjs/static-layer@3.1.6 beta
+npm dist-tag add @midwayjs/axios@3.1.5 latest
+npm dist-tag add @midwayjs/axios@3.1.6 beta
+npm dist-tag add @midwayjs/bootstrap@3.1.5 latest
+npm dist-tag add @midwayjs/bootstrap@3.1.6 beta
+npm dist-tag add @midwayjs/cache@3.1.5 latest
+npm dist-tag add @midwayjs/cache@3.1.6 beta
+npm dist-tag add @midwayjs/consul@3.1.5 latest
+npm dist-tag add @midwayjs/consul@3.1.6 beta
+npm dist-tag add @midwayjs/core@3.1.5 latest
+npm dist-tag add @midwayjs/core@3.1.6 beta
+npm dist-tag add @midwayjs/cos@3.1.5 latest
+npm dist-tag add @midwayjs/cos@3.1.6 beta
+npm dist-tag add @midwayjs/cross-domain@3.1.5 latest
+npm dist-tag add @midwayjs/cross-domain@3.1.6 beta
+npm dist-tag add @midwayjs/decorator@2.14.0 latest
+npm dist-tag add @midwayjs/decorator@3.1.6 beta
+npm dist-tag add @midwayjs/express-session@3.1.5 latest
+npm dist-tag add @midwayjs/express-session@3.1.6 beta
+npm dist-tag add @midwayjs/faas@3.1.5 latest
+npm dist-tag add @midwayjs/faas@3.1.6 beta
+npm dist-tag add @midwayjs/grpc@3.1.5 latest
+npm dist-tag add @midwayjs/grpc@3.1.6 beta
+npm dist-tag add @midwayjs/i18n@3.1.5 latest
+npm dist-tag add @midwayjs/i18n@3.1.6 beta
+npm dist-tag add @midwayjs/info@3.1.5 latest
+npm dist-tag add @midwayjs/info@3.1.6 beta
+npm dist-tag add @midwayjs/jwt@2.14.7 latest
+npm dist-tag add @midwayjs/jwt@3.1.6 beta
+npm dist-tag add @midwayjs/mock@3.1.5 latest
+npm dist-tag add @midwayjs/mock@3.1.6 beta
+npm dist-tag add @midwayjs/mongoose@3.1.5 latest
+npm dist-tag add @midwayjs/mongoose@3.1.6 beta
+npm dist-tag add @midwayjs/orm@2.14.7 latest
+npm dist-tag add @midwayjs/orm@3.1.6 beta
+npm dist-tag add @midwayjs/oss@3.1.5 latest
+npm dist-tag add @midwayjs/oss@3.1.6 beta
+npm dist-tag add @midwayjs/otel@3.1.5 latest
+npm dist-tag add @midwayjs/otel@3.1.6 beta
+npm dist-tag add @midwayjs/passport@2.14.7 latest
+npm dist-tag add @midwayjs/passport@3.1.6 beta
+npm dist-tag add @midwayjs/process-agent@3.1.5 latest
+npm dist-tag add @midwayjs/process-agent@3.1.6 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@2.14.7 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.1.6 beta
+npm dist-tag add @midwayjs/prometheus@3.1.5 latest
+npm dist-tag add @midwayjs/prometheus@3.1.6 beta
+npm dist-tag add @midwayjs/rabbitmq@3.1.5 latest
+npm dist-tag add @midwayjs/rabbitmq@3.1.6 beta
+npm dist-tag add @midwayjs/redis@3.1.5 latest
+npm dist-tag add @midwayjs/redis@3.1.6 beta
+npm dist-tag add @midwayjs/security@3.1.5 latest
+npm dist-tag add @midwayjs/security@3.1.6 beta
+npm dist-tag add @midwayjs/sequelize@3.1.5 latest
+npm dist-tag add @midwayjs/sequelize@3.1.6 beta
+npm dist-tag add @midwayjs/session@3.1.5 latest
+npm dist-tag add @midwayjs/session@3.1.6 beta
+npm dist-tag add @midwayjs/socketio@3.1.5 latest
+npm dist-tag add @midwayjs/socketio@3.1.6 beta
+npm dist-tag add @midwayjs/static-file@3.1.5 latest
+npm dist-tag add @midwayjs/static-file@3.1.6 beta
+npm dist-tag add @midwayjs/swagger@3.1.5 latest
+npm dist-tag add @midwayjs/swagger@3.1.6 beta
+npm dist-tag add @midwayjs/tablestore@3.1.5 latest
+npm dist-tag add @midwayjs/tablestore@3.1.6 beta
+npm dist-tag add @midwayjs/task@3.1.5 latest
+npm dist-tag add @midwayjs/task@3.1.6 beta
+npm dist-tag add @midwayjs/typegoose@3.1.5 latest
+npm dist-tag add @midwayjs/typegoose@3.1.6 beta
+npm dist-tag add @midwayjs/upload@3.1.5 latest
+npm dist-tag add @midwayjs/upload@3.1.6 beta
+npm dist-tag add @midwayjs/validate@3.1.5 latest
+npm dist-tag add @midwayjs/validate@3.1.6 beta
+npm dist-tag add @midwayjs/version@2.14.7 latest
+npm dist-tag add @midwayjs/version@3.1.6 beta
+npm dist-tag add @midwayjs/view-ejs@3.1.5 latest
+npm dist-tag add @midwayjs/view-ejs@3.1.6 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.1.5 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.1.6 beta
+npm dist-tag add @midwayjs/view@3.1.5 latest
+npm dist-tag add @midwayjs/view@3.1.6 beta
+npm dist-tag add @midwayjs/express@3.1.5 latest
+npm dist-tag add @midwayjs/express@3.1.6 beta
+npm dist-tag add @midwayjs/koa@3.1.5 latest
+npm dist-tag add @midwayjs/koa@3.1.6 beta
+npm dist-tag add @midwayjs/web@2.14.7 latest
+npm dist-tag add @midwayjs/web@3.1.6 beta
+npm dist-tag add @midwayjs/ws@3.1.5 latest
+npm dist-tag add @midwayjs/ws@3.1.6 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.14.7 latest
+tnpm dist-tag add @midwayjs/egg-layer@3.1.6 beta
+tnpm dist-tag add @midwayjs/express-layer@3.1.5 latest
+tnpm dist-tag add @midwayjs/express-layer@3.1.6 beta
+tnpm dist-tag add @midwayjs/faas-typings@3.1.5 latest
+tnpm dist-tag add @midwayjs/faas-typings@3.1.6 beta
+tnpm dist-tag add @midwayjs/koa-layer@3.1.5 latest
+tnpm dist-tag add @midwayjs/koa-layer@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-app@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-aws-starter@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-aws-starter@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-scf-starter@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-scf-starter@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@3.1.6 beta
+tnpm dist-tag add @midwayjs/serverless-worker-starter@3.1.5 latest
+tnpm dist-tag add @midwayjs/serverless-worker-starter@3.1.6 beta
+tnpm dist-tag add @midwayjs/static-layer@3.1.5 latest
+tnpm dist-tag add @midwayjs/static-layer@3.1.6 beta
+tnpm dist-tag add @midwayjs/axios@3.1.5 latest
+tnpm dist-tag add @midwayjs/axios@3.1.6 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.1.5 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.1.6 beta
+tnpm dist-tag add @midwayjs/cache@3.1.5 latest
+tnpm dist-tag add @midwayjs/cache@3.1.6 beta
+tnpm dist-tag add @midwayjs/consul@3.1.5 latest
+tnpm dist-tag add @midwayjs/consul@3.1.6 beta
+tnpm dist-tag add @midwayjs/core@3.1.5 latest
+tnpm dist-tag add @midwayjs/core@3.1.6 beta
+tnpm dist-tag add @midwayjs/cos@3.1.5 latest
+tnpm dist-tag add @midwayjs/cos@3.1.6 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.1.5 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.1.6 beta
+tnpm dist-tag add @midwayjs/decorator@2.14.0 latest
+tnpm dist-tag add @midwayjs/decorator@3.1.6 beta
+tnpm dist-tag add @midwayjs/express-session@3.1.5 latest
+tnpm dist-tag add @midwayjs/express-session@3.1.6 beta
+tnpm dist-tag add @midwayjs/faas@3.1.5 latest
+tnpm dist-tag add @midwayjs/faas@3.1.6 beta
+tnpm dist-tag add @midwayjs/grpc@3.1.5 latest
+tnpm dist-tag add @midwayjs/grpc@3.1.6 beta
+tnpm dist-tag add @midwayjs/i18n@3.1.5 latest
+tnpm dist-tag add @midwayjs/i18n@3.1.6 beta
+tnpm dist-tag add @midwayjs/info@3.1.5 latest
+tnpm dist-tag add @midwayjs/info@3.1.6 beta
+tnpm dist-tag add @midwayjs/jwt@2.14.7 latest
+tnpm dist-tag add @midwayjs/jwt@3.1.6 beta
+tnpm dist-tag add @midwayjs/mock@3.1.5 latest
+tnpm dist-tag add @midwayjs/mock@3.1.6 beta
+tnpm dist-tag add @midwayjs/mongoose@3.1.5 latest
+tnpm dist-tag add @midwayjs/mongoose@3.1.6 beta
+tnpm dist-tag add @midwayjs/orm@2.14.7 latest
+tnpm dist-tag add @midwayjs/orm@3.1.6 beta
+tnpm dist-tag add @midwayjs/oss@3.1.5 latest
+tnpm dist-tag add @midwayjs/oss@3.1.6 beta
+tnpm dist-tag add @midwayjs/otel@3.1.5 latest
+tnpm dist-tag add @midwayjs/otel@3.1.6 beta
+tnpm dist-tag add @midwayjs/passport@2.14.7 latest
+tnpm dist-tag add @midwayjs/passport@3.1.6 beta
+tnpm dist-tag add @midwayjs/process-agent@3.1.5 latest
+tnpm dist-tag add @midwayjs/process-agent@3.1.6 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.14.7 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.1.6 beta
+tnpm dist-tag add @midwayjs/prometheus@3.1.5 latest
+tnpm dist-tag add @midwayjs/prometheus@3.1.6 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.1.5 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.1.6 beta
+tnpm dist-tag add @midwayjs/redis@3.1.5 latest
+tnpm dist-tag add @midwayjs/redis@3.1.6 beta
+tnpm dist-tag add @midwayjs/security@3.1.5 latest
+tnpm dist-tag add @midwayjs/security@3.1.6 beta
+tnpm dist-tag add @midwayjs/sequelize@3.1.5 latest
+tnpm dist-tag add @midwayjs/sequelize@3.1.6 beta
+tnpm dist-tag add @midwayjs/session@3.1.5 latest
+tnpm dist-tag add @midwayjs/session@3.1.6 beta
+tnpm dist-tag add @midwayjs/socketio@3.1.5 latest
+tnpm dist-tag add @midwayjs/socketio@3.1.6 beta
+tnpm dist-tag add @midwayjs/static-file@3.1.5 latest
+tnpm dist-tag add @midwayjs/static-file@3.1.6 beta
+tnpm dist-tag add @midwayjs/swagger@3.1.5 latest
+tnpm dist-tag add @midwayjs/swagger@3.1.6 beta
+tnpm dist-tag add @midwayjs/tablestore@3.1.5 latest
+tnpm dist-tag add @midwayjs/tablestore@3.1.6 beta
+tnpm dist-tag add @midwayjs/task@3.1.5 latest
+tnpm dist-tag add @midwayjs/task@3.1.6 beta
+tnpm dist-tag add @midwayjs/typegoose@3.1.5 latest
+tnpm dist-tag add @midwayjs/typegoose@3.1.6 beta
+tnpm dist-tag add @midwayjs/upload@3.1.5 latest
+tnpm dist-tag add @midwayjs/upload@3.1.6 beta
+tnpm dist-tag add @midwayjs/validate@3.1.5 latest
+tnpm dist-tag add @midwayjs/validate@3.1.6 beta
+tnpm dist-tag add @midwayjs/version@2.14.7 latest
+tnpm dist-tag add @midwayjs/version@3.1.6 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.1.5 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.1.6 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.1.5 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.1.6 beta
+tnpm dist-tag add @midwayjs/view@3.1.5 latest
+tnpm dist-tag add @midwayjs/view@3.1.6 beta
+tnpm dist-tag add @midwayjs/express@3.1.5 latest
+tnpm dist-tag add @midwayjs/express@3.1.6 beta
+tnpm dist-tag add @midwayjs/koa@3.1.5 latest
+tnpm dist-tag add @midwayjs/koa@3.1.6 beta
+tnpm dist-tag add @midwayjs/web@2.14.7 latest
+tnpm dist-tag add @midwayjs/web@3.1.6 beta
+tnpm dist-tag add @midwayjs/ws@3.1.5 latest
+tnpm dist-tag add @midwayjs/ws@3.1.6 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.14.7 => 3.1.6
+#  - @midwayjs/express-layer: 3.1.5 => 3.1.6
+#  - @midwayjs/faas-typings: 3.1.5 => 3.1.6
+#  - @midwayjs/koa-layer: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-app: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-aws-starter: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-fc-starter: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-fc-trigger: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-http-parser: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-scf-starter: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-scf-trigger: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-vercel-starter: 3.1.5 => 3.1.6
+#  - @midwayjs/serverless-worker-starter: 3.1.5 => 3.1.6
+#  - @midwayjs/static-layer: 3.1.5 => 3.1.6
+#  - @midwayjs/axios: 3.1.5 => 3.1.6
+#  - @midwayjs/bootstrap: 3.1.5 => 3.1.6
+#  - @midwayjs/cache: 3.1.5 => 3.1.6
+#  - @midwayjs/consul: 3.1.5 => 3.1.6
+#  - @midwayjs/core: 3.1.5 => 3.1.6
+#  - @midwayjs/cos: 3.1.5 => 3.1.6
+#  - @midwayjs/cross-domain: 3.1.5 => 3.1.6
+#  - @midwayjs/decorator: 2.14.0 => 3.1.6
+#  - @midwayjs/express-session: 3.1.5 => 3.1.6
+#  - @midwayjs/faas: 3.1.5 => 3.1.6
+#  - @midwayjs/grpc: 3.1.5 => 3.1.6
+#  - @midwayjs/i18n: 3.1.5 => 3.1.6
+#  - @midwayjs/info: 3.1.5 => 3.1.6
+#  - @midwayjs/jwt: 2.14.7 => 3.1.6
+#  - @midwayjs/mock: 3.1.5 => 3.1.6
+#  - @midwayjs/mongoose: 3.1.5 => 3.1.6
+#  - @midwayjs/orm: 2.14.7 => 3.1.6
+#  - @midwayjs/oss: 3.1.5 => 3.1.6
+#  - @midwayjs/otel: 3.1.5 => 3.1.6
+#  - @midwayjs/passport: 2.14.7 => 3.1.6
+#  - @midwayjs/process-agent: 3.1.5 => 3.1.6
+#  - @midwayjs/prometheus-socket-io: 2.14.7 => 3.1.6
+#  - @midwayjs/prometheus: 3.1.5 => 3.1.6
+#  - @midwayjs/rabbitmq: 3.1.5 => 3.1.6
+#  - @midwayjs/redis: 3.1.5 => 3.1.6
+#  - @midwayjs/security: 3.1.5 => 3.1.6
+#  - @midwayjs/sequelize: 3.1.5 => 3.1.6
+#  - @midwayjs/session: 3.1.5 => 3.1.6
+#  - @midwayjs/socketio: 3.1.5 => 3.1.6
+#  - @midwayjs/static-file: 3.1.5 => 3.1.6
+#  - @midwayjs/swagger: 3.1.5 => 3.1.6
+#  - @midwayjs/tablestore: 3.1.5 => 3.1.6
+#  - @midwayjs/task: 3.1.5 => 3.1.6
+#  - @midwayjs/typegoose: 3.1.5 => 3.1.6
+#  - @midwayjs/upload: 3.1.5 => 3.1.6
+#  - @midwayjs/validate: 3.1.5 => 3.1.6
+#  - @midwayjs/version: 2.14.7 => 3.1.6
+#  - @midwayjs/view-ejs: 3.1.5 => 3.1.6
+#  - @midwayjs/view-nunjucks: 3.1.5 => 3.1.6
+#  - @midwayjs/view: 3.1.5 => 3.1.6
+#  - @midwayjs/express: 3.1.5 => 3.1.6
+#  - @midwayjs/koa: 3.1.5 => 3.1.6
+#  - @midwayjs/web: 2.14.7 => 3.1.6
+#  - @midwayjs/ws: 3.1.5 => 3.1.6