+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.6](https://github.com/midwayjs/midway/compare/v3.1.5...v3.1.6) (2022-03-21)
+
+
+### Bug Fixes
+
+* otel test ([#1837](https://github.com/midwayjs/midway/issues/1837)) ([36fa998](https://github.com/midwayjs/midway/commit/36fa998c077551e3ce9e0f0fda89d8bf1a3f0be7))
+
+
+
+
+
 ## [3.1.5](https://github.com/midwayjs/midway/compare/v3.1.4...v3.1.5) (2022-03-18)